refactor(adapters): move summary, dependencies, license_compliance to sections/ (#428)
Move three render functions from section.rs into dedicated files under sections/:
- sections/summary.rs: render_summary → render()
- sections/dependencies.rs: render_dependencies → render() with extracted render_component_row helper
- sections/license_compliance.rs: render_license_compliance → render()

Update sections/mod.rs to declare the three new submodules and update call sites
in markdown_formatter/mod.rs. Only render_resolution_guide remains in section.rs.

Closes #428
