+++ src/adapters/outbound/formatters/markdown_formatter/section.rs
@@ -1,314 +1,5 @@
-use crate::application::read_models::{
-    ComponentView, DependencyView, LicenseComplianceView, ResolutionGuideView,
-    UpgradeRecommendationView, VulnerabilityReportView,
-};
+use crate::application::read_models::{ResolutionGuideView, UpgradeRecommendationView};
 use crate::i18n::Messages;
-use std::collections::{HashMap, HashSet};
-
-/// Renders the executive summary section
-pub(super) fn render_summary(
-    messages: &'static Messages,
-    output: &mut String,
-    components: &[ComponentView],
-    vulnerabilities: Option<&VulnerabilityReportView>,
-    license_compliance: Option<&LicenseComplianceView>,
-) {
-    output.push_str(messages.section_summary);
-    output.push_str("\n\n");
-
-    // Table header
-    output.push_str(&format!(
-        "| {} | {} | {} |\n",
-        messages.col_item, messages.col_count, messages.col_status
-    ));
-    output.push_str(&super::table::make_separator(&[
-        messages.col_item,
-        messages.col_count,
-        messages.col_status,
-    ]));
-
-    // Package count rows
-    let direct_count = components.iter().filter(|c| c.is_direct_dependency).count();
-    let transitive_count = components
-        .iter()
-        .filter(|c| !c.is_direct_dependency)
-        .count();
-    output.push_str(&format!(
-        "| {} | {} | ✅ |\n",
-        messages.label_direct_deps, direct_count
-    ));
-    output.push_str(&format!(
-        "| {} | {} | ✅ |\n",
-        messages.label_transitive_deps, transitive_count
-    ));
-
-    // Vulnerability rows
-    let mut has_critical = false;
-    let mut has_warning = false;
-    if let Some(vuln_report) = vulnerabilities {
-        let counts = vuln_report.counts_by_severity();
-        let critical_status = if counts.critical > 0 {
-            has_critical = true;
-            "❌"
-        } else {
-            "✅"
-        };
-        let high_status = if counts.high > 0 {
-            has_warning = true;
-            "⚠️"
-        } else {
-            "✅"
-        };
-        let medium_status = if counts.medium > 0 {
-            has_warning = true;
-            "⚠️"
-        } else {
-            "✅"
-        };
-        let low_status = if counts.low > 0 {
-            has_warning = true;
-            "⚠️"
-        } else {
-            "✅"
-        };
-        output.push_str(&format!(
-            "| {} | {} | {} |\n",
-            messages.label_vuln_critical, counts.critical, critical_status
-        ));
-        output.push_str(&format!(
-            "| {} | {} | {} |\n",
-            messages.label_vuln_high, counts.high, high_status
-        ));
-        output.push_str(&format!(
-            "| {} | {} | {} |\n",
-            messages.label_vuln_medium, counts.medium, medium_status
-        ));
-        output.push_str(&format!(
-            "| {} | {} | {} |\n",
-            messages.label_vuln_low, counts.low, low_status
-        ));
-    } else {
-        output.push_str(&format!("\n{}\n", messages.label_vuln_check_skipped));
-    }
-
-    // License violations row
-    let violation_count = license_compliance
-        .map(|lc| lc.summary.violation_count)
-        .unwrap_or(0);
-    let license_status = if violation_count > 0 {
-        has_critical = true;
-        "❌"
-    } else {
-        "✅"
-    };
-    output.push_str(&format!(
-        "| {} | {} | {} |\n",
-        messages.label_license_violations, violation_count, license_status
-    ));
-
-    // Overall line
-    output.push('\n');
-    let overall = if has_critical {
-        messages.overall_action_required
-    } else if has_warning {
-        messages.overall_attention_recommended
-    } else {
-        messages.overall_no_issues
-    };
-    output.push_str(overall);
-    output.push_str("\n\n");
-}
-
-/// Renders the dependencies section
-pub(super) fn render_dependencies(
-    messages: &'static Messages,
-    verified_packages: Option<&HashSet<String>>,
-    output: &mut String,
-    deps: &DependencyView,
-    components: &[ComponentView],
-) {
-    // Create component lookup map by bom_ref
-    let component_map: HashMap<&str, &ComponentView> =
-        components.iter().map(|c| (c.bom_ref.as_str(), c)).collect();
-
-    // Direct Dependencies section
-    output.push_str(messages.section_direct_deps);
-    output.push_str("\n\n");
-    output.push_str(messages.desc_direct_deps);
-    output.push_str("\n\n");
-
-    if !deps.direct.is_empty() {
-        output.push_str(&super::table::table_header(messages));
-        output.push_str(&super::table::table_separator(messages));
-
-        for bom_ref in &deps.direct {
-            if let Some(component) = component_map.get(bom_ref.as_str()) {
-                let license = component
-                    .license
-                    .as_ref()
-                    .map(|l| l.spdx_id.as_deref().unwrap_or(l.name.as_str()))
-                    .unwrap_or("N/A");
-                let description = component.description.as_deref().unwrap_or("");
-
-                output.push_str(&format!(
-                    "| {} | {} | {} | {} |\n",
-                    super::links::format_package_name(&component.name, verified_packages),
-                    super::table::escape_markdown_table_cell(&component.version),
-                    super::table::escape_markdown_table_cell(license),
-                    super::table::escape_markdown_table_cell(description)
-                ));
-            }
-        }
-        output.push('\n');
-    } else {
-        output.push_str(messages.label_no_direct_deps);
-        output.push_str("\n\n");
-    }
-
-    // Transitive Dependencies section
-    output.push_str(messages.section_transitive_deps);
-    output.push_str("\n\n");
-    output.push_str(messages.desc_transitive_deps);
-    output.push_str("\n\n");
-
-    if !deps.transitive.is_empty() {
-        for direct_ref in &deps.direct {
-            if let Some(trans_deps) = deps.transitive.get(direct_ref) {
-                if trans_deps.is_empty() {
-                    continue;
-                }
-
-                // Get direct dependency name for header
-                let parent_name = component_map
-                    .get(direct_ref.as_str())
-                    .map(|c| c.name.as_str())
-                    .unwrap_or(direct_ref);
-
-                output.push_str(&Messages::format(messages.deps_for_header, &[parent_name]));
-                output.push_str("\n\n");
-                output.push_str(&super::table::table_header(messages));
-                output.push_str(&super::table::table_separator(messages));
-
-                for trans_ref in trans_deps {
-                    if let Some(component) = component_map.get(trans_ref.as_str()) {
-                        let license = component
-                            .license
-                            .as_ref()
-                            .map(|l| l.spdx_id.as_deref().unwrap_or(l.name.as_str()))
-                            .unwrap_or("N/A");
-                        let description = component.description.as_deref().unwrap_or("");
-
-                        output.push_str(&format!(
-                            "| {} | {} | {} | {} |\n",
-                            super::links::format_package_name(&component.name, verified_packages),
-                            super::table::escape_markdown_table_cell(&component.version),
-                            super::table::escape_markdown_table_cell(license),
-                            super::table::escape_markdown_table_cell(description)
-                        ));
-                    }
-                }
-                output.push('\n');
-            }
-        }
-    } else {
-        output.push_str(messages.label_no_transitive_deps);
-        output.push_str("\n\n");
-    }
-}
-
-/// Renders the license compliance section
-pub(super) fn render_license_compliance(
-    messages: &'static Messages,
-    output: &mut String,
-    compliance: &LicenseComplianceView,
-) {
-    output.push('\n');
-    output.push_str(messages.section_license_compliance);
-    output.push_str("\n\n");
-
-    // Summary
-    if compliance.has_violations {
-        let unit = if compliance.summary.violation_count == 1 {
-            messages.label_license_violation_singular
-        } else {
-            messages.label_license_violation_plural
-        };
-        output.push_str(&format!(
-            "**{} {}**\n\n",
-            compliance.summary.violation_count, unit,
-        ));
-    } else {
-        output.push_str(messages.label_no_license_violations);
-        output.push_str("\n\n");
-    }
-
-    // Violations table
-    if !compliance.violations.is_empty() {
-        output.push_str(messages.section_violations);
-        output.push_str("\n\n");
-        output.push_str(&format!(
-            "| {} | {} | {} | {} | {} |\n",
-            messages.col_package,
-            messages.col_version,
-            messages.col_license,
-            messages.col_reason,
-            messages.col_matched_pattern,
-        ));
-        output.push_str(&super::table::make_separator(&[
-            messages.col_package,
-            messages.col_version,
-            messages.col_license,
-            messages.col_reason,
-            messages.col_matched_pattern,
-        ]));
-
-        for v in &compliance.violations {
-            output.push_str(&format!(
-                "| {} | {} | {} | {} | {} |\n",
-                super::table::escape_markdown_table_cell(&v.package_name),
-                super::table::escape_markdown_table_cell(&v.package_version),
-                super::table::escape_markdown_table_cell(&v.license),
-                super::table::escape_markdown_table_cell(&v.reason),
-                v.matched_pattern.as_deref().unwrap_or("-"),
-            ));
-        }
-        output.push('\n');
-    }
-
-    // Warnings table
-    if !compliance.warnings.is_empty() {
-        let warning_count = compliance.summary.warning_count;
-        let pkg_word = if warning_count == 1 {
-            messages.label_package_singular
-        } else {
-            messages.label_package_plural
-        };
-        output.push_str(messages.section_warnings);
-        output.push_str("\n\n");
-        output.push_str(&Messages::format(
-            messages.warn_unknown_license_packages,
-            &[&warning_count.to_string(), pkg_word],
-        ));
-        output.push_str("\n\n");
-        output.push_str(&format!(
-            "| {} | {} |\n",
-            messages.col_package, messages.col_version,
-        ));
-        output.push_str(&super::table::make_separator(&[
-            messages.col_package,
-            messages.col_version,
-        ]));
-
-        for w in &compliance.warnings {
-            output.push_str(&format!(
-                "| {} | {} |\n",
-                super::table::escape_markdown_table_cell(&w.package_name),
-                super::table::escape_markdown_table_cell(&w.package_version),
-            ));
-        }
-        output.push('\n');
-    }
-}
 
 /// Renders the resolution guide section
 pub(super) fn render_resolution_guide(
