+++ src/adapters/outbound/formatters/markdown_formatter/sections/mod.rs
@@ -1,3 +1,6 @@
 pub(super) mod components;
+pub(super) mod dependencies;
 pub(super) mod header;
+pub(super) mod license_compliance;
 pub(super) mod section_renderer;
+pub(super) mod summary;
